User POST and GET Implemented
- Created methods for /users endpoint:
- > POST - Registration (moved from auth folder)
- > GET - Display all the users on the platform, with the count of their friends, posts, likes, and comments
